requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
flask==2.3.2
flask-cors==4.0.0
requests==2.31.0